naive version of finding close balls

+++ types.h
@@ -44,7 +44,7 @@ enum OBJECT
 };
 
 
-const int NUMBER_OF_BALLS = 11;
+const size_t NUMBER_OF_BALLS = 11;
 
 enum STATE
 {
